// src/core/evm_chains/types.ts
export type GasfeeResult = {
  slow: number;
  average: number;
  fast: number;
};

export type EvmTxData = {
  transaction_hash: string;
  from: string;
  to: string;
  amount: number;
  gasFee: number;
  block_number: number;
  block_hash: string;
  gasPrice: string;
  nonce: number;
};

export type EvmTransferParams = {
  recipient: string;
  amount: number;
  fee?: number;
  memo?: string;
};